Split parser behavior declarations
